// core/admin-dashboard/store/user/action-types.ts
export enum ActionTypes {
  GET_USERS = "GET_USERS",
  GET_USER = "GET_USER",
  CREATE_USER = "CREATE_USER",
  UPDATE_USER = "UPDATE_USER",
  DELETE_USER = "DELETE_USER",
  UPLOAD_USER_AVATAR = "UPLOAD_USER_AVATAR",
  HARD_DELETE_USER = "HARD_DELETE_USER",
  BLOCK_USER_COMMENT = "BLOCK_USER_COMMENT",
  UNBLOCK_USER_COMMENT = "UNBLOCK_USER_COMMENT",
  UPDATE_USER_PASSWORD = "UPDATE_USER_PASSWORD",
  RESTORE_USER = "RESTORE_USER",
  GET_USER_ANALYTICS = "GET_USER_ANALYTICS",
  RESET_USER_LOGIN_FAILED_TIMES = "RESET_USER_LOGIN_FAILED_TIMES",
  BATCH_UPLOAD_USERS = "BATCH_UPLOAD_USERS",
}